feat(react-components): Shared modal
