+++ libs/portals/admin/ids-admin/src/components/forms/PublishEnvironment/PublishEnvironment.tsx
@@ -5,7 +5,7 @@ import {
   Text,
   toast,
 } from '@island.is/island-ui/core'
-import { Modal } from '../../Modal/Modal'
+import { Modal } from '@island.is/react/components'
 import { m } from '../../../lib/messages'
 import { Form, useActionData, useNavigate, useParams } from 'react-router-dom'
 import { useLocale } from '@island.is/localization'
@@ -73,6 +73,7 @@ export default function PublishEnvironment() {
         environment: publishData?.toEnvironment,
       })}
       onClose={cancel}
+      closeButtonLabel={formatMessage(m.closeModal)}
     >
       <Form method="post">
         <Box paddingTop={3}>
