+++ libs/react/components/src/index.ts
@@ -1,1 +1,2 @@
 export * from './components/LoadingScreen/LoadingScreen'
+export * from './components/Modal/Modal'
